feat: refactor the shell component
